requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
emoji==2.8.0
filelock==3.18.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.4.0
google-api-core==2.25.1
google-auth==2.40.3
google-generativeai==0.3.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.62.3
huggingface-hub==0.34.3
idna==3.10
importlib-metadata==6.11.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
narwhals==2.0.1
networkx==3.3
numpy==1.26.4
packaging==23.2
pandas==2.3.1
pillow==10.4.0
proto-plus==1.26.1
protobuf==4.25.8
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
rich==13.9.4
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.3
setuptools==70.2.0
six==1.17.0
smmap==5.0.2
streamlit==1.29.0
sympy==1.13.3
tenacity==8.5.0
tokenizers==0.21.4
toml==0.10.2
torch==2.7.1+cpu
tornado==6.5.1
tqdm==4.67.1
transformers==4.55.0
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
validators==0.35.0
watchdog==6.0.0
zipp==3.23.0
matplotlib